Task2/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
pillow
requests
gunicorn==20.1.0